Arreglos para muestra funcional

+++ src/Controlador/ModificarUsuario_Eventos.java
@@ -211,7 +211,8 @@ public class ModificarUsuario_Eventos
         {
             try
             {
-                Integer.parseInt(usuario.getTelefono());
+//                Integer.parseInt(usuario.getTelefono());
+                Long.parseLong(usuario.getTelefono());
             }
             catch (NumberFormatException ex)
             {
@@ -224,7 +225,8 @@ public class ModificarUsuario_Eventos
         {
             try
             {
-                Integer.parseInt(usuario.getCelular());
+//                Integer.parseInt(usuario.getCelular());
+                Long.parseLong(usuario.getCelular());
             }
             catch (NumberFormatException ex)
             {
